refactor: extract formatting logic from assembler.rs into format.rs; fix docs links
- memory-core/src/context/dag/format.rs: new file with formatting/display logic (format_for_prompt, format_compact, format_full, format_optimized, reduction_percentage, node_type_name) extracted from assembler.rs to bring it under 500 LOC
- memory-core/src/context/dag/assembler.rs: delegate formatting to super::format::* with identical public method signatures
- memory-core/src/context/dag/mod.rs: add pub mod format
- .agents/skills/agent-coordination/SKILL.md: fix typo skills-agants.md -> skills-agents.md
- README.md: remove broken coverage.svg badge reference (file does not exist)
