+++ memory-core/src/context/dag/mod.rs
@@ -33,6 +33,7 @@
 
 pub mod assembler;
 pub mod edge;
+pub mod format;
 pub mod node;
 pub mod state;
